+++ src/main/kotlin/by/mrvanish97/kbnsext/plugin/BeansPluginBundle.kt
@@ -1,0 +1,31 @@
+/*******************************************************************************
+ * Copyright (c) 2021 mrvanish97 [and others]
+ *
+ * This program and the accompanying materials are made
+ * available under the terms of the Eclipse Public License 2.0
+ * which is available at https://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ ******************************************************************************/
+
+package by.mrvanish97.kbnsext.plugin
+
+import com.intellij.AbstractBundle
+import org.jetbrains.annotations.NonNls
+import org.jetbrains.annotations.PropertyKey
+
+@NonNls
+private const val BUNDLE = "messages.BeansPluginBundle"
+
+object BeansPluginBundle : AbstractBundle(BUNDLE) {
+
+    @Suppress("SpreadOperator")
+    @JvmStatic
+    fun message(@PropertyKey(resourceBundle = BUNDLE) key: String, vararg params: Any) =
+        getMessage(key, *params)
+
+    @Suppress("SpreadOperator")
+    @JvmStatic
+    fun messagePointer(@PropertyKey(resourceBundle = BUNDLE) key: String, vararg params: Any) =
+        getLazyMessage(key, *params)
+}