+++ src/main/kotlin/icons/BeansPluginIcons.kt
@@ -1,0 +1,20 @@
+/*******************************************************************************
+ * Copyright (c) 2021 mrvanish97 [and others]
+ *
+ * This program and the accompanying materials are made
+ * available under the terms of the Eclipse Public License 2.0
+ * which is available at https://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ ******************************************************************************/
+
+package icons
+
+import com.intellij.openapi.util.IconLoader
+
+object BeansPluginIcons {
+
+  @JvmField
+  val kotlinBeanScriptIcon = IconLoader.getIcon("/icons/kotlinBeansScriptIcon.svg", javaClass)
+
+}